Merging r242574: ------------------------------------------------------------------------ r242574 | compnerd | 2015-07-17 14:26:41 -0700 (Fri, 17 Jul 2015) | 8 lines
basic: default to MSVC on Windows

The "armv7-windows", "i686-windows", and "x86_64-windows" targets should be
equivalent to the MSVC environment.  This was previously discussed when the
triples for Windows werw canonicalised.  Im not sure how this was overlooked.
This fixes the emission of non-COFF formats on Windows.

Thanks to ki9a for reporting this issue over IRC!
------------------------------------------------------------------------

git-svn-id: https://llvm.org/svn/llvm-project/cfe/branches/release_37@242597 91177308-0d34-0410-b5e6-96231b3b80d8